+++ NEWS.md
@@ -1,3 +1,15 @@
+# hintr 0.1.25
+
+* Queue calibration with higher priority than model run
+
+# hintr 0.1.24
+
+* Add endpoints to enable async calibration
+   * /calibrate/submit - queue a model calibration
+   * /calibrate/status - get status of a calibration run
+   * /calibrate/result - get result of a calibration
+* Add calibration options back in and remove from run options
+
 # hintr 0.1.23
 
 * Update to naomi v2.1.11 (branch pin naomi@naomisup-ticket-11).